requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==2.3.3
gunicorn==21.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.3
packaging==24.2
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.1
requests==2.31.0
six==1.17.0
slack-bolt==1.18.0
slack-sdk==3.23.0
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
